source/core/src/test/com/csse3200/game/components/player/InventoryComponentTest.java
package com.csse3200.game.components.player;

import com.csse3200.game.extensions.GameExtension;
import org.junit.jupiter.api.Test;
import org.junit.jupiter.api.extension.ExtendWith;
import static org.junit.jupiter.api.Assertions.assertEquals;
import static org.junit.jupiter.api.Assertions.assertTrue;
import static org.junit.jupiter.api.Assertions.assertFalse;

@ExtendWith(GameExtension.class)
class InventoryComponentTest {
  @Test
  void shouldSetSize() {
    InventoryComponent inventory = new InventoryComponent(2);
    assertEquals(2, inventory.getSize());

    inventory.setSize(1);
    assertEquals(1, inventory.getSize());

    try {
        inventory.setSize(-1);
    } catch (IllegalArgumentException e) {
        assertEquals("Invalid size parameter. Must be an integer > 0", e.getMessage());
    }
  }

  /*
  @Test
  void shouldCheckHasGold() {
    InventoryComponent inventory = new InventoryComponent(150);
    assertTrue(inventory.hasGold(100));
    assertFalse(inventory.hasGold(200));
  }

  @Test
  void shouldAddGold() {
    InventoryComponent inventory = new InventoryComponent(100);
    inventory.addGold(-500);
    assertEquals(0, inventory.getGold());

    inventory.addGold(100);
    inventory.addGold(-20);
    assertEquals(80, inventory.getGold());
  }
  */

}
